--- api/schoolbus/src/main/java/com/schoolbus/model/Person.java
@@ -1,81 +1,0 @@
-package com.schoolbus.model;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class Person {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    private String name;
-    private int age;
-    private String cpf;
-    private String rg;
-    private String address;
-
-    public Person() {
-
-    }
-
-    public Person(String name, int age, String cpf, String rg, String address) {
-
-        super();
-        setName(name);
-        setAge(age);
-        setCpf(cpf);
-        setRg(rg);
-        setAddress(address);
-        
-    }
-
-    public Integer getId() {
-
-        return id;
-    }
-
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setCpf(String cpf) {
-        this.cpf = cpf;
-    }
-
-    public String getCpf() {
-        return cpf;
-    }
-
-    public void setRg(String rg) {
-        this.rg = rg;
-    }
-
-    public String getRg() {
-        return rg;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-}
