--- api/schoolbus/src/main/java/com/schoolbus/repository/PersonRepository.java
@@ -1,9 +1,0 @@
-package com.schoolbus.repository;
-
-import com.schoolbus.model.Student;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PersonRepository extends JpaRepository<Student, Integer> {
-
-}
