+++ api/schoolbus/src/main/java/com/schoolbus/resources/StudentResources.java
@@ -9,7 +9,6 @@ import com.schoolbus.repository.StudentRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,22 +16,14 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 
 
-@Controller
+@RestController
 public class StudentResources {
 
     @Autowired
     private StudentRepository repository;
-
-    // @RequestMapping("/")
-    // public ModelAndView Hello() {
-        
-    //     ModelAndView modelAndView = new ModelAndView();
-    //     modelAndView.setViewName("index.html");
-    //     return modelAndView;
-        
-    // }
 
     @GetMapping("/students")
     public List<Student> getAllStudents() {
@@ -43,7 +34,7 @@ public class StudentResources {
     }
 
     @PostMapping("/students")
-    public ResponseEntity saveStudent(@RequestBody Student student) {
+    public ResponseEntity<?> saveStudent(@RequestBody Student student) {
         
         Student savedStudent = repository.save(student);
 
@@ -75,12 +66,11 @@ public class StudentResources {
 
     @DeleteMapping("/students/{id}")
     public void deleteStudent(@PathVariable int id) {
-
+        
         Optional<Student> studentFound = repository.findById(id);
 
         if (studentFound.isPresent()) {
             repository.deleteById(id);
         }
     }
-    
 }
